resolve circular dependency issues
