+++ src/services/WalletService/wallets/helpers.ts
@@ -1,10 +1,4 @@
-import TransportWebUSB from '@ledgerhq/hw-transport-webusb';
-
-import { DPath, WalletId } from '@types';
-
-import { LedgerU2F, LedgerUSB } from './ledger';
-import Trezor from './Trezor';
-
+import { DPath } from '@types';
 /**
  * Get the full derivation path with address index.
  *
@@ -16,16 +10,4 @@ export const getFullPath = (derivationPath: DPath, addrIndex: number): string =>
     return derivationPath.value.replace('addrIndex', addrIndex.toString());
   }
   return `${derivationPath.value}/${addrIndex}`;
-};
-
-export const selectWallet = async (walletId: WalletId) => {
-  switch (walletId) {
-    default:
-    case WalletId.LEDGER_NANO_S_NEW: {
-      const isWebUSBSupported = await TransportWebUSB.isSupported().catch(() => false);
-      return isWebUSBSupported ? new LedgerUSB() : new LedgerU2F();
-    }
-    case WalletId.TREZOR_NEW:
-      return new Trezor();
-  }
 };
